#31 - create call to create detail screen and some improvements
