+++ app/src/main/java/com/codandotv/streamplayerapp/navigation/NavigationGraph.kt
@@ -13,15 +13,17 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import com.codandotv.streamplayerapp.core_navigation.bottomnavigation.StreamPlayerBottomNavigation
 import com.codandotv.streamplayerapp.core_navigation.routes.BottomNavRoutes
-import com.codandotv.streamplayerapp.core_navigation.routes.SplashRoutes
-import com.codandotv.streamplayerapp.feature_list_streams.presentation.navigation.listStreamsNavGraph
+import com.codandotv.streamplayerapp.core_navigation.routes.Routes
+import com.codandotv.streamplayerapp.feature_list_streams.detail.presentation.navigation.detailStreamNavGraph
+import com.codandotv.streamplayerapp.feature_list_streams.list.presentation.navigation.listStreamsNavGraph
 import com.codandotv.streamplayerapp.splah.presentation.navigation.splashNavGraph
 
 @Composable
 fun NavigationGraph(navController: NavHostController) {
-    NavHost(navController = navController, startDestination = SplashRoutes.Splash) {
+    NavHost(navController = navController, startDestination = Routes.Splash) {
         splashNavGraph(navController = navController)
         listStreamsNavGraph(navController = navController)
+        detailStreamNavGraph(navController = navController)
         temporaryFun(BottomNavRoutes.GAMES, navController)
         temporaryFun(BottomNavRoutes.NEWS, navController)
         temporaryFun(BottomNavRoutes.SCENES, navController)
