+++ app/src/main/java/com/codandotv/streamplayerapp/splah/presentation/navigation/SplashNavigation.kt
@@ -4,11 +4,11 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
 import com.codandotv.streamplayerapp.core_navigation.routes.BottomNavRoutes
-import com.codandotv.streamplayerapp.core_navigation.routes.SplashRoutes
+import com.codandotv.streamplayerapp.core_navigation.routes.Routes
 import com.codandotv.streamplayerapp.splah.presentation.screens.SplashScreen
 
 fun NavGraphBuilder.splashNavGraph(navController: NavHostController) {
-    composable(SplashRoutes.Splash) {
+    composable(Routes.Splash) {
         SplashScreen(
             onAnimationFinished = { navController.navigate(BottomNavRoutes.HOME) }
         )
