+++ buildSrc/src/main/java/Config.kt
@@ -9,7 +9,7 @@ object Config {
     const val packageName = "com.codandotv.streamplayerapp."
 
     object BuildField {
-        const val host_debug = "\"https://api.themoviedb.org/\""
-        const val host_release = "\"https://api.themoviedb.org/\""
+        const val host_debug = "\"https://api.themoviedb.org/3/\""
+        const val host_release = "\"https://api.themoviedb.org/3/\""
     }
 }
