+++ buildSrc/src/main/java/Versions.kt
@@ -14,6 +14,7 @@ object Versions {
     const val androidCoreKtx = "1.7.0"
     const val support_version = "1.4.1"
     const val material_version = "1.5.0"
+    const val dynamic_animation = "1.0.0"
     const val retrofit_version = "2.9.0"
     const val moshi_version = "1.14.0"
     const val okhttp_version = "4.10.0"
