+++ core-navigation/src/main/java/com/codandotv/streamplayerapp/core_navigation/routes/Routes.kt
@@ -1,0 +1,13 @@
+package com.codandotv.streamplayerapp.core_navigation.routes
+
+import com.codandotv.streamplayerapp.core_navigation.routes.Routes.PARAM.ID
+
+object Routes {
+    const val DETAIL = "DetailList/"
+    const val DETAIL_COMPLETE = "${DETAIL}{${ID}}"
+    const val Splash = "splash"
+
+    object PARAM {
+        const val ID = "id"
+    }
+}
