--- core-navigation/src/main/java/com/codandotv/streamplayerapp/core_navigation/routes/SplashRoutes.kt
@@ -1,5 +1,0 @@
-package com.codandotv.streamplayerapp.core_navigation.routes
-
-object SplashRoutes {
-    const val Splash = "splash"
-}
