+++ core-networking/src/main/java/com/codandotv/streamplayerapp/core_networking/Url.kt
@@ -1,0 +1,6 @@
+package com.codandotv.streamplayerapp.core_networking
+
+object Url {
+    const val IMAGE_URL_SIZE_300 = "https://image.tmdb.org/t/p/w300/"
+    const val IMAGE_URL_SIZE_500 = "https://image.tmdb.org/t/p/w500/"
+}
