+++ core-networking/src/main/java/com/codandotv/streamplayerapp/core_networking/di/NetworkModule.kt
@@ -1,8 +1,7 @@
 package com.codandotv.streamplayerapp.core_networking.di
 
 import com.codandotv.streamplayerapp.core_networking.BuildConfig
-import com.codandotv.streamplayerapp.core_networking.handleError.coroutines.NetworkResponseAdapter
-import com.codandotv.streamplayerapp.core_networking.handleError.coroutines.NetworkResponseAdapterFactory
+import com.codandotv.streamplayerapp.core_networking.coroutines.NetworkResponseAdapterFactory
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import okhttp3.Interceptor
