--- feature-list-streams/src/main/java/com/codandotv/streamplayerapp/feature_list_streams/data/model/GenresResponse.kt
@@ -1,10 +1,0 @@
-package com.codandotv.streamplayerapp.feature_list_streams.data.model
-
-data class GenreResponse(
-    val id: Long,
-    val name: String
-)
-
-data class GenresResponse(
-    val genres: List<GenreResponse>
-)
