--- feature-list-streams/src/main/java/com/codandotv/streamplayerapp/feature_list_streams/data/model/ListStreamResponse.kt
@@ -1,10 +1,0 @@
-package com.codandotv.streamplayerapp.feature_list_streams.data.model
-
-data class StreamResponse(
-    val title : String,
-    val overview : String,
-    val poster_path: String,
-)
-data class ListStreamResponse(
-    val results: List<StreamResponse>
-)
