+++ feature-list-streams/src/main/java/com/codandotv/streamplayerapp/feature_list_streams/detail/data/DetailStreamService.kt
@@ -1,0 +1,11 @@
+package com.codandotv.streamplayerapp.feature_list_streams.detail.data
+
+import com.codandotv.streamplayerapp.core_networking.handleError.NetworkResponse
+import com.codandotv.streamplayerapp.feature_list_streams.detail.data.model.DetailStreamResponse
+import retrofit2.http.GET
+import retrofit2.http.Path
+
+interface DetailStreamService {
+    @GET("movie/{movie_id}")
+    suspend fun getMovie(@Path("movie_id") movieId: String) : NetworkResponse<DetailStreamResponse>
+}
