+++ feature-list-streams/src/main/java/com/codandotv/streamplayerapp/feature_list_streams/detail/data/model/DetailStreamResponse.kt
@@ -1,0 +1,9 @@
+package com.codandotv.streamplayerapp.feature_list_streams.detail.data.model
+
+data class DetailStreamResponse(
+    val id : String,
+    val title : String,
+    val overview : String,
+    val tagline : String,
+    val backdrop_path : String,
+)
