+++ feature-list-streams/src/main/java/com/codandotv/streamplayerapp/feature_list_streams/detail/domain/DetailStream.kt
@@ -1,0 +1,9 @@
+package com.codandotv.streamplayerapp.feature_list_streams.detail.domain
+
+data class DetailStream(
+    val id : String,
+    val title : String,
+    val overview : String,
+    val tagline : String,
+    val imagePath : String
+)
