+++ feature-list-streams/src/main/java/com/codandotv/streamplayerapp/feature_list_streams/detail/presentation/screens/DetailStreamsUIState.kt
@@ -1,0 +1,14 @@
+package com.codandotv.streamplayerapp.feature_list_streams.detail.presentation.screens
+
+import com.codandotv.streamplayerapp.feature_list_streams.detail.domain.DetailStream
+import com.codandotv.streamplayerapp.feature_list_streams.list.presentation.widgets.StreamsCardContent
+
+data class CarouselData(
+    val categoryName: String,
+    val cards: List<StreamsCardContent>
+)
+
+data class DetailStreamsUIState(
+    val detailStream: DetailStream? = null,
+    val isLoading: Boolean,
+)
