--- feature-list-streams/src/main/java/com/codandotv/streamplayerapp/feature_list_streams/domain/ListStreamAnalytics.kt
@@ -1,6 +1,0 @@
-package com.codandotv.streamplayerapp.feature_list_streams.domain
-
-interface ListStreamAnalytics{}
-
-class ListStreamAnalyticsImpl : ListStreamAnalytics {
-}
