--- feature-list-streams/src/main/java/com/codandotv/streamplayerapp/feature_list_streams/domain/model/ListStream.kt
@@ -1,11 +1,0 @@
-package com.codandotv.streamplayerapp.feature_list_streams.domain.model
-
-data class Stream(
-    val name : String,
-    val description : String,
-    val posterPathUrl: String,
-)
-data class ListStream(
-    val categoryName: String,
-    val streams: List<Stream>
-)
