+++ feature-list-streams/src/main/java/com/codandotv/streamplayerapp/feature_list_streams/list/data/model/GenresResponse.kt
@@ -1,0 +1,10 @@
+package com.codandotv.streamplayerapp.feature_list_streams.list.data.model
+
+data class GenreResponse(
+    val id: Long,
+    val name: String
+)
+
+data class GenresResponse(
+    val genres: List<GenreResponse>
+)
