+++ feature-list-streams/src/main/java/com/codandotv/streamplayerapp/feature_list_streams/list/data/model/ListStreamResponse.kt
@@ -1,0 +1,11 @@
+package com.codandotv.streamplayerapp.feature_list_streams.list.data.model
+
+data class StreamResponse(
+    val id : String,
+    val title : String,
+    val overview : String,
+    val poster_path: String,
+)
+data class ListStreamResponse(
+    val results: List<StreamResponse>
+)
