+++ feature-list-streams/src/main/java/com/codandotv/streamplayerapp/feature_list_streams/list/domain/ListStreamAnalytics.kt
@@ -1,0 +1,6 @@
+package com.codandotv.streamplayerapp.feature_list_streams.list.domain
+
+interface ListStreamAnalytics{}
+
+class ListStreamAnalyticsImpl : ListStreamAnalytics {
+}
