--- feature-list-streams/src/main/java/com/codandotv/streamplayerapp/feature_list_streams/presentation/screens/ListStreamsScreen.kt
@@ -1,81 +1,0 @@
-package com.codandotv.streamplayerapp.feature_list_streams.presentation.screens
-
-import android.annotation.SuppressLint
-import androidx.compose.foundation.ScrollState
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import androidx.navigation.NavController
-import androidx.navigation.compose.rememberNavController
-import com.codandotv.streamplayerapp.core_navigation.bottomnavigation.StreamPlayerBottomNavigation
-import com.codandotv.streamplayerapp.core_shared_ui.theme.ThemePreviews
-import com.codandotv.streamplayerapp.feature_list_streams.presentation.widgets.StreamsCarousel
-import org.koin.androidx.compose.koinViewModel
-
-@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ListStreamsScreen(
-    navController: NavController,
-    viewModel: ListStreamViewModel = koinViewModel()
-) {
-
-    val uiState = remember {
-        viewModel.uiState
-    }
-
-    Scaffold(
-        bottomBar = {
-            StreamPlayerBottomNavigation(navController = navController)
-        }
-    ) { _ ->
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(MaterialTheme.colorScheme.background)
-        ) {
-            if (uiState.value.isLoading) {
-                CircularProgressIndicator(
-                    modifier = Modifier.align(
-                        Alignment.Center
-                    )
-                )
-            } else {
-                Column(
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .align(Alignment.TopCenter)
-                        .verticalScroll(ScrollState(0))
-                ) {
-                    uiState.value.carousels.forEach {
-                        StreamsCarousel(
-                            title = it.categoryName,
-                            contentList = it.cards
-                        )
-                    }
-
-                    Spacer(modifier = Modifier.height(12.dp))
-                }
-            }
-        }
-    }
-}
-
-@ThemePreviews
-@Composable
-fun ListStreamsScreenPreview() {
-    ListStreamsScreen(navController = rememberNavController())
-}
