--- feature-list-streams/src/main/java/com/codandotv/streamplayerapp/feature_list_streams/presentation/widgets/StreamsCarousel.kt
@@ -1,47 +1,0 @@
-package com.codandotv.streamplayerapp.feature_list_streams.presentation.widgets
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-
-@Composable
-fun StreamsCarousel(title: String, contentList: List<StreamsCardContent>, modifier: Modifier = Modifier) {
-    Column(modifier = modifier) {
-        Text(
-            title,
-            style = MaterialTheme.typography.headlineMedium.copy(
-                color = Color.White,
-                fontWeight = FontWeight.Bold,
-                fontSize = 20.sp
-            )
-        )
-
-        Spacer(modifier = Modifier.size(8.dp))
-
-        LazyRow(modifier = Modifier.fillMaxWidth()) {
-            items(contentList.size) {
-                StreamsCard(contentList[it])
-            }
-        }
-    }
-}
-
-@Composable
-@Preview
-fun StreamsCarouselPreview() {
-    StreamsCarousel(
-        title = "Ação",
-        contentList = emptyList()
-    )
-}
